+++ packages/web-components/src/components/date-picker/index.ts
@@ -5,7 +5,11 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import './components/date-picker/date-picker';
-import './components/date-picker/date-picker-input';
-import './components/date-picker/date-picker-input-skeleton';
-import './components/date-picker/calendar-renderer';
+// Web Components implementation
+import './components/date-picker';
+import './components/date-picker-input';
+import './components/date-picker-input-skeleton';
+import './components/calendar-renderer';
+
+// Export state machine from primitives package
+export * from '@carbon-labs/primitives/date-picker';